Added removing from favorites budgets button

+++ views/budget_utils/remove_budget.php
@@ -8,7 +8,7 @@ if (!isset($_SESSION['logged'])) {
 
 header('Content-Type: application/json');
 
-function removeBudget($budgetID) {
+function removeBudget($userID, $budgetID) {
     require_once '../../connect.php';
     $conn = new mysqli($host, $db_user, $db_password, $db_name);
 
@@ -18,23 +18,39 @@ function removeBudget($budgetID) {
         exit();
     }
 
-    $sql = "DELETE FROM budgets WHERE Budget_id = ?";
-    $stmt = $conn->prepare($sql);
-    $stmt->bind_param("i", $budgetID);
-
-    if ($stmt->execute()) {
-        echo json_encode(["success" => true, "message" => "Budżet został pomyślnie usunięty!"]);
-    } else {
-        error_log("SQL error: " . $stmt->error);
-        echo json_encode(["success" => false, "message" => "Błąd podczas usuwania budżetu."]);
+    // Usuwanie transakcji z budzetu
+    $sql_remove_transactions = "DELETE FROM transactions WHERE User_id = ? AND Budget_id = ?";
+    $stmt_remove_transactions = $conn->prepare($sql_remove_transactions);
+    $stmt_remove_transactions->bind_param("ii", $userID, $budgetID);
+    if($stmt_remove_transactions->execute()){
+        $sql = "DELETE FROM budgets WHERE User_id = ? AND Budget_id = ?";
+        $stmt = $conn->prepare($sql);
+        $stmt->bind_param("ii", $userID, $budgetID);
+    
+        if ($stmt->execute()) {
+            echo json_encode(["success" => true, "message" => "Budżet został pomyślnie usunięty!"]);
+        } else {
+            echo json_encode(["success" => false, "message" => "Błąd podczas usuwania budżetu."]);
+        }
+    
+        $stmt->close();
+    }
+    else{
+        echo json_encode(["success" => false, "message" => "Błąd podczas usuwania transakcji."]);
     }
 
-    $stmt->close();
     $conn->close();
 }
 
 if ($_SERVER['REQUEST_METHOD'] == 'POST') {
     $budgetID = $_POST['budget_id'] ?? null;
+    $userID = $_SESSION['id'] ?? null;
+
+    if(!$userID) {
+        error_log('Brak ID użytkownika w sesji');
+        echo json_encode(["success" => false, "message" => "Brak ID użytkownika."]);
+        exit();
+    }
 
     if (!$budgetID) {
         error_log('Brak ID budżetu w żądaniu POST');
@@ -42,8 +58,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
         exit();
     }
 
-    error_log('Otrzymano żądanie usunięcia budżetu z ID: ' . $budgetID);
-    removeBudget($budgetID);
+    removeBudget($userID,$budgetID);
 }
 
 ?>
